[VOL-4123] Fix no flows being discarded by ofAgent
This commit fixes an issue in ofAgent where it fails to send
stats to ONOS if there are no flows/ no ports/ no port desc.  This
situation arises if the rw-core does not persist flows and is
restarted.

This fix was tested with a rw-core that does not persist flows,
meters and groups.  The rw-core was restarted after 512 ONUs were
already activated and had their flows pushed.  After the rw-core
retart, with this fix, the flows were pushed from ONOS.

Change-Id: I10ee627e4153134ec5a5d12d1dcd67f64213e83d